Moving ReplaceThreadRunWithStart to cleanup package
