+++ rewrite-java-tck/src/main/kotlin/org/openrewrite/java/cleanup/ReplaceThreadRunWithThreadStartTest.kt
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java
+package org.openrewrite.java.cleanup
 
 import org.junit.jupiter.api.Test
 import org.openrewrite.Recipe
+import org.openrewrite.java.JavaRecipeTest
 
 interface ReplaceThreadRunWithThreadStartTest : JavaRecipeTest {
     override val recipe: Recipe
