+++ rewrite-java/src/main/java/org/openrewrite/java/cleanup/ReplaceThreadRunWithThreadStart.java
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java;
+package org.openrewrite.java.cleanup;
+
+import org.openrewrite.java.ChangeMethodInvocation;
 
 import java.util.Collections;
 import java.util.Set;
